Reland "[lldb][AArch64] Linux corefile support for SME"
This reverts commit 3fa503582315f23f187a019f026c5fce59b3f3d8.

m_sve_state was not initialised which (I'm guessing) meant that
it could potentially be a value that matched a real SVE state.
Then we'd be acting as if we're streaming mode, for example,
without ever having the data required to back that up.

By sheer luck this only turn up on x86, AArch64 and ARM were fine.
It is UB regardless.