[mlir][Vector] Fix doc generation (#67341)
The format is
`add_mlir_doc(<table gen file> <output md name> <output dir> <command>)`

Using the Vector.td file does not work as it does not include
VectorOps.td (so the docs are empty). If the output name is not
VectorOps.md, the mlir-www site does not find the ops (and instead
overwrites it with an empty template).